requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.0
numpy==1.18.0
pyyaml==5.2
